Move CQN specific type to their own package
