+++ types/types.go
@@ -1,0 +1,21 @@
+package types
+
+type CqnData struct {
+	SchemaTableName string
+	TableOperation  CqnOpCode
+	RowChanges      RowChanges
+}
+
+const (
+	CqnAllRows CqnOpCode = 1 << iota
+	CqnInsert
+	CqnUpdate
+	CqnDelete
+	CqnAlter
+	CqnDrop
+	CqnUnexpected
+)
+
+type RowId string
+type CqnOpCode uint32
+type RowChanges map[RowId]CqnOpCode
